+++ linuxdoom-1.10/p_sight.cpp
@@ -310,9 +310,9 @@ P_CheckSight
     // First check for trivial rejection.
 
     // Determine subsector entries in REJECT table.
-    s1 = (t1->subsector->sector - sectors.data());
-    s2 = (t2->subsector->sector - sectors.data());
-    pnum = s1*sectors.size() + s2;
+    s1 = static_cast<int>(t1->subsector->sector - sectors.data());
+    s2 = static_cast<int>(t2->subsector->sector - sectors.data());
+    pnum = static_cast<int>(s1*sectors.size() + s2);
     bytenum = pnum>>3;
     bitnum = 1 << (pnum&7);
